mise a jour propel rev 2327
git-svn-id: http://svn.sylogix.org/gepiphp/trunk@7343 e073496e-d129-0410-8f7b-841bfd467322

+++ orm/propel/collection/PropelObjectCollection.php
@@ -16,17 +16,23 @@
  */
 class PropelObjectCollection extends PropelCollection
 {
-	
+
 	/**
 	 * Save all the elements in the collection
+	 *
+	 * @param     PropelPDO  $con
 	 */
 	public function save($con = null)
 	{
+		if (!method_exists($this->getModel(), 'save')) {
+			throw new PropelException('Cannot save objects on a read-only model');
+		}
 		if (null === $con) {
 			$con = $this->getConnection(Propel::CONNECTION_WRITE);
 		}
 		$con->beginTransaction();
 		try {
+			/** @var $element BaseObject */
 			foreach ($this as $element) {
 				$element->save($con);
 			}
@@ -36,17 +42,23 @@ class PropelObjectCollection extends PropelCollection
 			throw $e;
 		}
 	}
-	
+
 	/**
 	 * Delete all the elements in the collection
+	 *
+	 * @param     PropelPDO  $con
 	 */
 	public function delete($con = null)
 	{
+		if (!method_exists($this->getModel(), 'delete')) {
+			throw new PropelException('Cannot delete objects on a read-only model');
+		}
 		if (null === $con) {
 			$con = $this->getConnection(Propel::CONNECTION_WRITE);
 		}
 		$con->beginTransaction();
 		try {
+			/** @var $element BaseObject */
 			foreach ($this as $element) {
 				$element->delete($con);
 			}
@@ -60,16 +72,19 @@ class PropelObjectCollection extends PropelCollection
 	/**
 	 * Get an array of the primary keys of all the objects in the collection
 	 *
-	 * @return    array The list of the primary keys of the collection
+	 * @param     boolean  $usePrefix
+	 * @return    array  The list of the primary keys of the collection
 	 */
 	public function getPrimaryKeys($usePrefix = true)
 	{
 		$ret = array();
+
+		/** @var $obj BaseObject */
 		foreach ($this as $key => $obj) {
 			$key = $usePrefix ? ($this->getModel() . '_' . $key) : $key;
 			$ret[$key]= $obj->getPrimaryKey();
 		}
-		
+
 		return $ret;
 	}
 
@@ -78,31 +93,31 @@ class PropelObjectCollection extends PropelCollection
 	 * Each object is populated from an array and the result is stored
 	 * Does not empty the collection before adding the data from the array
 	 *
-	 * @param    array $arr
+	 * @param    array  $arr
 	 */
 	public function fromArray($arr)
 	{
 		$class = $this->getModel();
 		foreach ($arr as $element) {
+			/** @var $obj BaseObject */
 			$obj = new $class();
 			$obj->fromArray($element);
 			$this->append($obj);
 		}
 	}
-	
+
 	/**
 	 * Get an array representation of the collection
 	 * Each object is turned into an array and the result is returned
 	 *
 	 * @param     string  $keyColumn If null, the returned array uses an incremental index.
-	 *                    Otherwise, the array is indexed using the specified column
-	 * @param     boolean $usePrefix If true, the returned array prefixes keys 
-	 *                    with the model class name ('Article_0', 'Article_1', etc).
-	 * @param     string  $keyType (optional) One of the class type constants BasePeer::TYPE_PHPNAME, 
-	 *                    BasePeer::TYPE_STUDLYPHPNAME, BasePeer::TYPE_COLNAME, BasePeer::TYPE_FIELDNAME,
-	 *                    BasePeer::TYPE_NUM. Defaults to BasePeer::TYPE_PHPNAME.
-	 * @param     boolean $includeLazyLoadColumns (optional) Whether to include lazy loaded columns.
-	 *                    Defaults to TRUE.
+	 *                               Otherwise, the array is indexed using the specified column
+	 * @param     boolean $usePrefix If true, the returned array prefixes keys
+	 *                               with the model class name ('Article_0', 'Article_1', etc).
+	 * @param     string  $keyType   (optional) One of the class type constants BasePeer::TYPE_PHPNAME,
+	 *                               BasePeer::TYPE_STUDLYPHPNAME, BasePeer::TYPE_COLNAME, BasePeer::TYPE_FIELDNAME,
+	 *                               BasePeer::TYPE_NUM. Defaults to BasePeer::TYPE_PHPNAME.
+	 * @param     boolean $includeLazyLoadColumns (optional) Whether to include lazy loaded columns. Defaults to TRUE.
 	 * @param     array   $alreadyDumpedObjects List of objects to skip to avoid recursion
 	 *
 	 * <code>
@@ -122,46 +137,50 @@ class PropelObjectCollection extends PropelCollection
 	 *  'Book_1' => array('Id' => 456, 'Title' => 'Don Juan'),
 	 * )
 	 * </code>
+	 *
 	 * @return    array
 	 */
 	public function toArray($keyColumn = null, $usePrefix = false, $keyType = BasePeer::TYPE_PHPNAME, $includeLazyLoadColumns = true, $alreadyDumpedObjects = array())
 	{
 		$ret = array();
 		$keyGetterMethod = 'get' . $keyColumn;
+
+		/** @var $obj BaseObject */
 		foreach ($this as $key => $obj) {
 			$key = null === $keyColumn ? $key : $obj->$keyGetterMethod();
 			$key = $usePrefix ? ($this->getModel() . '_' . $key) : $key;
 			$ret[$key] = $obj->toArray($keyType, $includeLazyLoadColumns, $alreadyDumpedObjects, true);
 		}
-		
+
 		return $ret;
 	}
-	
+
 	/**
 	 * Get an array representation of the collection
 	 *
-	 * @param     string $keyColumn If null, the returned array uses an incremental index.
-	 *              Otherwise, the array is indexed using the specified column
-	 * @param     boolean $usePrefix If true, the returned array prefixes keys 
-	 *              with the model class name ('Article_0', 'Article_1', etc).
+	 * @param     string   $keyColumn  If null, the returned array uses an incremental index.
+	 *                                 Otherwise, the array is indexed using the specified column
+	 * @param     boolean  $usePrefix  If true, the returned array prefixes keys
+	 *                                 with the model class name ('Article_0', 'Article_1', etc).
 	 *
 	 * <code>
-	 * $bookCollection->getArrayCopy();
-	 * array(
-	 *  0 => $book0,
-	 *  1 => $book1,
-	 * )
-	 * $bookCollection->getArrayCopy('Id');
-	 * array(
-	 *  123 => $book0,
-	 *  456 => $book1,
-	 * )
-	 * $bookCollection->getArrayCopy(null, true);
-	 * array(
-	 *  'Book_0' => $book0,
-	 *  'Book_1' => $book1,
-	 * )
+	 *   $bookCollection->getArrayCopy();
+	 *   array(
+	 *    0 => $book0,
+	 *    1 => $book1,
+	 *   )
+	 *   $bookCollection->getArrayCopy('Id');
+	 *   array(
+	 *    123 => $book0,
+	 *    456 => $book1,
+	 *   )
+	 *   $bookCollection->getArrayCopy(null, true);
+	 *   array(
+	 *    'Book_0' => $book0,
+	 *    'Book_1' => $book1,
+	 *   )
 	 * </code>
+	 *
 	 * @return    array
 	 */
 	public function getArrayCopy($keyColumn = null, $usePrefix = false)
@@ -176,17 +195,21 @@ class PropelObjectCollection extends PropelCollection
 			$key = $usePrefix ? ($this->getModel() . '_' . $key) : $key;
 			$ret[$key] = $obj;
 		}
-		
+
 		return $ret;
 	}
-	
+
 	/**
 	 * Get an associative array representation of the collection
 	 * The first parameter specifies the column to be used for the key,
 	 * And the seconf for the value.
+	 *
 	 * <code>
-	 * $res = $coll->toKeyValue('Id', 'Name');
+	 *   $res = $coll->toKeyValue('Id', 'Name');
 	 * </code>
+	 *
+	 * @param     string  $keyColumn
+	 * @param     string  $valueColumn
 	 *
 	 * @return    array
 	 */
@@ -198,19 +221,19 @@ class PropelObjectCollection extends PropelCollection
 		foreach ($this as $obj) {
 			$ret[$obj->$keyGetterMethod()] = $obj->$valueGetterMethod();
 		}
-		
+
 		return $ret;
 	}
-	
+
 	/**
 	 * Makes an additional query to populate the objects related to the collection objects
 	 * by a certain relation
 	 *
-	 * @param     string    $relation Relation name (e.g. 'Book')
-	 * @param     Criteria  $criteria Optional Criteria object to filter the related object collection
-	 * @param     PropelPDO $con      Optional connection object
+	 * @param     string     $relation  Relation name (e.g. 'Book')
+	 * @param     Criteria   $criteria  Optional Criteria object to filter the related object collection
+	 * @param     PropelPDO  $con       Optional connection object
 	 *
-	 * @return    PropelObjectCollection the list of related objects
+	 * @return    PropelObjectCollection  The list of related objects
 	 */
 	public function populateRelation($relation, $criteria = null, $con = null)
 	{
@@ -221,12 +244,12 @@ class PropelObjectCollection extends PropelCollection
 		if ($this->isEmpty()) {
 			// save a useless query and return an empty collection
 			$coll = new PropelObjectCollection();
-			$coll->setModel($relationMap->getRightTable()->getPhpName());
+			$coll->setModel($relationMap->getRightTable()->getClassname());
 			return $coll;
 		}
 		$symRelationMap = $relationMap->getSymmetricalRelation();
-		
-		$query = PropelQuery::from($relationMap->getRightTable()->getPhpName());
+
+		$query = PropelQuery::from($relationMap->getRightTable()->getClassname());
 		if (null !== $criteria) {
 			$query->mergeWith($criteria);
 		}
@@ -236,9 +259,14 @@ class PropelObjectCollection extends PropelCollection
 			->$filterMethod($this)
 			->find($con);
 		if ($relationMap->getType() == RelationMap::ONE_TO_MANY) {
+			// initialize the embedded collections of the main objects
+			$relationName = $relationMap->getName();
+			foreach ($this as $mainObj) {
+				$mainObj->initRelation($relationName);
+			}
 			// associate the related objects to the main objects
 			$getMethod = 'get' . $symRelationMap->getName();
-			$addMethod = 'add' . $relationMap->getName();
+			$addMethod = 'add' . $relationName;
 			foreach ($relatedObjects as $object) {
 				$mainObj = $object->$getMethod();  // instance pool is used here to avoid a query
 				$mainObj->$addMethod($object);
@@ -249,10 +277,7 @@ class PropelObjectCollection extends PropelCollection
 		} else {
 			throw new PropelException('populateRelation() does not support this relation type');
 		}
-		
+
 		return $relatedObjects;
 	}
-
 }
-
-?>
